# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CommentCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "anchor": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor"
          },
          "line": {
            "exclusiveMinimum": 0.0,
            "title": "Line",
            "type": "integer"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "line",
          "text"
        ],
        "title": "CommentCreateRequest",
        "type": "object"
      },
      "CommentReplyRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "author_type": {
            "const": "human",
            "title": "Author Type",
            "type": "string"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "author_type"
        ],
        "title": "CommentReplyRequest",
        "type": "object"
      },
      "CommentResolveRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          }
        },
        "required": [
          "resolved"
        ],
        "title": "CommentResolveRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PrdCommentRead": {
        "additionalProperties": false,
        "properties": {
          "author_type": {
            "enum": [
              "human",
              "agent"
            ],
            "title": "Author Type",
            "type": "string"
          },
          "body": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "id": {
            "exclusiveMinimum": 0.0,
            "title": "Id",
            "type": "integer"
          },
          "line": {
            "exclusiveMinimum": 0.0,
            "title": "Line",
            "type": "integer"
          },
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "replies": {
            "items": {
              "$ref": "#/components/schemas/PrdCommentReplyRead"
            },
            "title": "Replies",
            "type": "array"
          },
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "path",
          "line",
          "author_type",
          "body",
          "resolved",
          "replies"
        ],
        "title": "PrdCommentRead",
        "type": "object"
      },
      "PrdCommentReplyRead": {
        "additionalProperties": false,
        "description": "A read-only reply nested under its authoritative Gitea comment.",
        "properties": {
          "author_type": {
            "enum": [
              "human",
              "agent"
            ],
            "title": "Author Type",
            "type": "string"
          },
          "body": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "id": {
            "exclusiveMinimum": 0.0,
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "author_type",
          "body"
        ],
        "title": "PrdCommentReplyRead",
        "type": "object"
      },
      "PrdCommentableLineRead": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "context",
              "addition"
            ],
            "title": "Kind",
            "type": "string"
          },
          "line": {
            "exclusiveMinimum": 0.0,
            "title": "Line",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "line",
          "kind",
          "text"
        ],
        "title": "PrdCommentableLineRead",
        "type": "object"
      },
      "PrdCommentableLinesRead": {
        "additionalProperties": false,
        "properties": {
          "commit_sha": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Commit Sha",
            "type": "string"
          },
          "filename": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PrdCommentableLineRead"
            },
            "title": "Lines",
            "type": "array"
          },
          "version": {
            "exclusiveMinimum": 0.0,
            "title": "Version",
            "type": "integer"
          },
          "wi": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Wi",
            "type": "string"
          }
        },
        "required": [
          "wi",
          "version",
          "filename",
          "commit_sha",
          "lines"
        ],
        "title": "PrdCommentableLinesRead",
        "type": "object"
      },
      "PrdDiffRead": {
        "additionalProperties": false,
        "description": "Full target-file PR patch, including deletion lines that cannot be annotated.",
        "properties": {
          "commit_sha": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Commit Sha",
            "type": "string"
          },
          "filename": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "patch": {
            "title": "Patch",
            "type": "string"
          },
          "version": {
            "exclusiveMinimum": 0.0,
            "title": "Version",
            "type": "integer"
          },
          "wi": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Wi",
            "type": "string"
          }
        },
        "required": [
          "wi",
          "version",
          "filename",
          "commit_sha",
          "patch"
        ],
        "title": "PrdDiffRead",
        "type": "object"
      },
      "PrdDocumentRead": {
        "additionalProperties": false,
        "properties": {
          "change_summary": {
            "anyOf": [
              {
                "maxLength": 8000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Summary"
          },
          "commit_sha": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha"
          },
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "filename": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "pr_number": {
            "exclusiveMinimum": 0.0,
            "title": "Pr Number",
            "type": "integer"
          },
          "version": {
            "exclusiveMinimum": 0.0,
            "title": "Version",
            "type": "integer"
          },
          "wi": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Wi",
            "type": "string"
          }
        },
        "required": [
          "wi",
          "version",
          "filename",
          "pr_number",
          "content"
        ],
        "title": "PrdDocumentRead",
        "type": "object"
      },
...
      "PublishReviewAccepted": {
        "additionalProperties": false,
        "description": "The frozen task identity returned before background publication.",
        "properties": {
          "base_version": {
            "exclusiveMinimum": 0.0,
            "title": "Base Version",
            "type": "integer"
          },
          "comment_count": {
            "exclusiveMinimum": 0.0,
            "title": "Comment Count",
            "type": "integer"
          },
          "task_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "base_version",
          "comment_count"
        ],
        "title": "PublishReviewAccepted",
        "type": "object"
      },
      "PublishReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          }
        },
        "title": "PublishReviewRequest",
        "type": "object"
      },
...
      "ReviewTaskRead": {
        "additionalProperties": false,
        "description": "Public task progress, intentionally excluding internal failure detail.",
        "properties": {
          "base_version": {
            "exclusiveMinimum": 0.0,
            "title": "Base Version",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "new_commit_sha": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Commit Sha"
          },
          "new_version": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Version"
          },
          "status": {
            "enum": [
              "pending",
              "processing",
              "done",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Task Id",
            "type": "string"
          },
          "wi": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Wi",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "wi",
          "status",
          "base_version"
        ],
        "title": "ReviewTaskRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/prd/{wi}": {
      "get": {
        "operationId": "latest_prd_prd__wi__get",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrdDocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Prd",
        "tags": [
          "prd-review"
        ]
      }
    },
    "/prd/{wi}/commentable-lines": {
      "get": {
        "operationId": "prd_commentable_lines_prd__wi__commentable_lines_get",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrdCommentableLinesRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prd Commentable Lines",
        "tags": [
          "prd-review"
        ]
      }
    },
    "/prd/{wi}/comments": {
      "get": {
        "operationId": "prd_comments_prd__wi__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrdCommentRead"
                  },
                  "title": "Response Prd Comments Prd  Wi  Comments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prd Comments",
        "tags": [
          "prd-review"
        ]
      },
      "post": {
        "operationId": "create_comment_prd__wi__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Comment",
        "tags": [
          "prd-review"
        ]
      }
    },
    "/prd/{wi}/comments/{comment_id}/reply": {
      "post": {
        "operationId": "reply_comment_prd__wi__comments__comment_id__reply_post",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentReplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reply Comment",
        "tags": [
          "prd-review"
        ]
      }
    },
    "/prd/{wi}/comments/{comment_id}/resolve": {
      "post": {
        "operationId": "resolve_comment_prd__wi__comments__comment_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Comment",
        "tags": [
          "prd-review"
        ]
      }
    },
    "/prd/{wi}/diff": {
      "get": {
        "operationId": "prd_diff_prd__wi__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrdDiffRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prd Diff",
        "tags": [
          "prd-review"
        ]
      }
    },
    "/prd/{wi}/reviews/publish": {
      "post": {
        "operationId": "publish_review_prd__wi__reviews_publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReviewAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Review",
        "tags": [
          "prd-review"
        ]
      }
    },
    "/prd/{wi}/v/{number}": {
      "get": {
        "operationId": "prd_version_prd__wi__v__number__get",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "title": "Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrdDocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prd Version",
        "tags": [
          "prd-review"
        ]
      }
    },
    "/prd/{wi}/versions": {
      "get": {
        "operationId": "prd_versions_prd__wi__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "wi",
            "required": true,
            "schema": {
              "title": "Wi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrdDocumentRead"
                  },
                  "title": "Response Prd Versions Prd  Wi  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prd Versions",
        "tags": [
          "prd-review"
        ]
      }
    },
...
    "/tasks/{task_id}": {
      "get": {
        "operationId": "review_task_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTaskRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Task",
        "tags": [
          "prd-review"
        ]
      }
    }
  }
}